05-clothingController: getClothing

+++ develop/src/main/java/cz/cvut/fel/nss/vestana/repo/ClothingArticleRepo.java
@@ -5,5 +5,5 @@ import org.springframework.data.repository.CrudRepository;
 import org.springframework.stereotype.Repository;
 
 @Repository
-public interface ClothingArticleRepo extends CrudRepository<ClothingArticle, Integer> {
+public interface ClothingArticleRepo extends CrudRepository<ClothingArticle, Long> {
 }
